+++ src/main/java/org/truffleruby/core/exception/CoreExceptions.java
@@ -181,8 +181,18 @@ public class CoreExceptions {
     }
 
     @TruffleBoundary
-    public RubyException argumentErrorUnknownKeyword(Object key, Node currentNode) {
-        return argumentError("unknown keyword: " + inspect(key), currentNode);
+    public RubyException argumentErrorUnknownKeywords(Object[] keys, Node currentNode) {
+        if (keys.length == 1) {
+            return argumentError("unknown keyword: " + inspect(keys[0]), currentNode);
+        }
+
+        final String[] names = new String[keys.length];
+
+        for (int i = 0; i < keys.length; i++) {
+            names[i] = inspect(keys[i]);
+        }
+
+        return argumentError("unknown keywords: " + String.join(", ", names), currentNode);
     }
 
     @TruffleBoundary
@@ -197,8 +207,18 @@ public class CoreExceptions {
     }
 
     @TruffleBoundary
-    public RubyException argumentErrorMissingKeyword(String name, Node currentNode) {
-        return argumentError(StringUtils.format("missing keyword: %s", name), currentNode);
+    public RubyException argumentErrorMissingKeywords(Object[] keys, Node currentNode) {
+        if (keys.length == 1) {
+            return argumentError("missing keyword: " + inspect(keys[0]), currentNode);
+        }
+
+        final String[] names = new String[keys.length];
+
+        for (int i = 0; i < keys.length; i++) {
+            names[i] = inspect(keys[i]);
+        }
+
+        return argumentError("missing keywords: " + String.join(", ", names), currentNode);
     }
 
     @TruffleBoundary