+++ src/main/java/org/truffleruby/parser/LoadArgumentsTranslator.java
@@ -27,7 +27,6 @@ import org.truffleruby.language.SourceIndexLength;
 import org.truffleruby.language.arguments.ArrayIsAtLeastAsLargeAsNode;
 import org.truffleruby.language.arguments.CheckNoKeywordArgumentsNode;
 import org.truffleruby.language.arguments.MissingArgumentBehavior;
-import org.truffleruby.language.arguments.MissingKeywordArgumentNode;
 import org.truffleruby.language.arguments.ReadKeywordArgumentNode;
 import org.truffleruby.language.arguments.ReadKeywordRestArgumentNode;
 import org.truffleruby.language.arguments.ReadOptionalArgumentNode;
@@ -273,9 +272,7 @@ public class LoadArgumentsTranslator extends Translator {
 
         final RubyNode defaultValue;
         if (asgnNode.getValueNode() instanceof RequiredKeywordArgumentValueParseNode) {
-            /* This isn't a true default value - it's a marker to say there isn't one. This actually makes sense; the
-             * semantic action of executing this node is to report an error, and we do the same thing. */
-            defaultValue = new MissingKeywordArgumentNode(name);
+            defaultValue = null;
         } else {
             defaultValue = translateNodeOrNil(sourceSection, asgnNode.getValueNode());
         }